frontend/requirements.txt
# Streamlit 프론트엔드 의존성
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
streamlit-camera-input-live>=0.2.0

# HTTP 클라이언트
requests>=2.31.0
httpx>=0.25.0

# 이미지 처리
Pillow>=10.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 유틸리티
pydantic>=2.4.0 